% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.three}
\alias{as.three}
\alias{as.three,igraph-method}
\title{generate 3d graph}
\usage{
\S4method{as.three}{igraph}(object, type = "base", bgColor = "0xffffff",
  nodeSize = 5, edgeColor = "0x000000", edgeWidth = 5, fontSize = 20,
  fontColor = "0x000000", fontOffset = c(x = 5, y = 5, z = 5))
}
\arguments{
\item{object}{the object}

\item{type}{display mode}

\item{bgColor}{a hex value}

\item{nodeSize}{size of nods}

\item{edgeColor}{a hex value}

\item{edgeWidth}{defaults to 5}

\item{fontSize}{defaults to 20}

\item{fontColor}{a hex value}

\item{fontOffset}{a list}
}
\description{
generate 3d graph with three.js
}
\examples{
\dontrun{
bt17merkel <- partition("PLPRTXT", list(text_lp="17", text_speaker="Angela Merkel", text_type="speech"), tf="word")
bt17merkelColl <- collocations(bt17merkel, pAttribute="word", mc=TRUE)
bt17merkelCollTrimmed <- trim(bt17merkelColl, cutoff=list(ll=100 ))
iMerkel <- asIgraph(bt17merkelCollTrimmed)
iMerkelComm <- enrich(iMerkel, community=list(method="fastgreedy", weights=FALSE))
iMerkel2d <- enrich(iMerkelComm, layout="fruchterman.reingold", dim=2)
iMerkel3d <- enrich(iMerkelComm, layout="kamada.kawai", dim=3)
iMerkel3d <- rescale(iMerkel3d, -200, 200)
iMerkelJs <- as.three(iMerkel3d, bgColor="0xcccccc", fontSize=12, fontColor="0x000000", nodeSize=4, edgeColor="0xeeeeee", edgeWidth=1, fontOffset=c(x=10,y=10,z=10))
show(iMerkelJs)
DataTablesR:::browse(iMerkelJs)
}
}

