% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's Algorithm}
\usage{
dijkstra(wiki_graph, init_node)
}
\arguments{
\item{init_node}{A numeric scalar representing the initial node, which must exist in the graph.}

\item{graph}{A data.frame with three columns (v1, v2, w) representing the edges 
of the graph and their corresponding weights. v1 and v2 are the nodes 
and w is the weight of the edge.}
}
\value{
A vector representing the shortest path from the initial node to every other node.
}
\description{
This function implements Dijkstra's algorithm to calculate the shortest path 
from the initial node to all other nodes in the graph.
}
\details{
The function will throw an error if the graph structure is incorrect or 
         if the initial node is not in the graph.
}
\examples{
wiki_graph <-data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
          v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
          w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))
dijkstra(wiki_graph, 1)
}
\references{
See [Dijkstra's Algorithm](https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm)
}
