% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Calculate Euclidean distance}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{The coordinates of the first point (numeric value)}

\item{b}{The coordinates of the second point (numeric value)}
}
\value{
Returns the Euclidean distance (numeric value) between two points.
}
\description{
This function implements the Euclidean algorithm to calculate the distance between two points.
}
\examples{
euclidean(1, 2)  # Calculate the distance between points (1, 2) and (0, 0)
}
\references{
See [Wikipedia](https://en.wikipedia.org/wiki/Euclidean_distance).
}
