% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{polmineR.three-package}
\alias{polmineR.three-package}
\title{polmineR.three-package}
\description{
3D graph visualization with three.js
}
\details{
A story to be told
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\references{
http://polmine.sowi.uni-due.de
}
\keyword{package}

