% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/three-package.R
\docType{package}
\name{three-package}
\alias{three-package}
\title{three-package}
\description{
3D visualization with three.js
}
\details{
A story to be told
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\references{
http://polmine.sowi.uni-due.de
}
\keyword{package}

