% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm for GCD}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric value representing the first integer.}

\item{b}{A numeric value representing the second integer.}
}
\value{
The GCD of the two input integers.
}
\description{
This function calculates the Greatest Common Divisor (GCD) of two integers using the Euclidean algorithm.
}
\details{
The Euclidean algorithm works by repeatedly applying the division algorithm. The process continues
until the remainder is zero, at which point the divisor is the GCD of the two numbers.
}
\examples{
euclidean(252, 105)
euclidean(123612, 13892347912)

}
\references{
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{Euclidean Algorithm - Wikipedia}
}
