% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{rescale}
\alias{rescale}
\alias{rescale,igraph-method}
\title{rescale coordinates of an object}
\usage{
\S4method{rescale}{igraph}(object, min, max)
}
\arguments{
\item{object}{object}

\item{min}{min}

\item{max}{max}
}
\description{
bla
}

