% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{three-class}
\alias{three-class}
\title{three class for preparing three.js scripts}
\description{
to be displayed with a browser
}
\section{Slots}{

\describe{
\item{\code{js}}{character, the JavaScript code}

\item{\code{json}}{a list with character strings}

\item{\code{type}}{the display mode}
}}

