\name{Three-package}
\alias{Three-package}
\alias{Three}
\docType{package}
\title{
\packageTitle{Three}
}
\description{
\packageDescription{Three}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{Three}

Maintainer: \packageMaintainer{Three}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = Three)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{Three}}
%\section{Documentation Index}{\packageIndices{Three}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
