\name{Three-package}
\alias{Three-package}
\alias{Three}
\docType{package}
\title{
\packageTitle{Three}
}
\description{
\packageDescription{Three}
}

\author{
\packageAuthor{Three}

Maintainer: \packageMaintainer{Three}
}

\keyword{package}

%% Uncomment below to imitate parts of library(help = Three)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{Three}}
%\section{Documentation Index}{\packageIndices{Three}}


