% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{light,three-method}
\alias{light}
\alias{light,three-method}
\alias{lines,three-method}
\alias{points,three-method}
\alias{show,three-method}
\alias{text,three-method}
\alias{three}
\title{create a three-object}
\usage{
\S4method{light}{three}(x, pointLight = list(x = 0, y = 0, z = 0))

\S4method{lines}{three}(x, from, to, color = "0x000000", lwd = 5)

\S4method{points}{three}(x, coords, size = 5, color = "0xcccccc")

\S4method{text}{three}(x, .data, color = "0x000000", size = 15,
  offset = c(x = 5, y = 5, z = 5))

\S4method{show}{three}(object)

three(type = "base", bgColor = "0x000000")
}
\arguments{
\item{x}{a three object}

\item{pointLight}{a list with the x,y,z-coordinates of a point light}

\item{from}{data frame with coordinates of starting points of lines}

\item{to}{data frame with coordinates of end points of lines}

\item{color}{color of line / point etc.}

\item{lwd}{line width, defaults to 5}

\item{coords}{coordinates of points}

\item{size}{point size}

\item{.data}{a data.frame with three columns providing x/y/z coordinates, and rownames providing text}

\item{offset}{offset of the text}

\item{object}{a three object}

\item{type}{character vector providing display mode ("base", "anaglyph", "stereo")}

\item{bgColor}{the background color}
}
\description{
create a three-object
}

