% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three-class.R, R/light-method.R, R/lines-method.R, R/points-method.R, R/text-method.R, R/three.R
\docType{class}
\name{three-class}
\alias{light}
\alias{light,three-method}
\alias{lines,three-method}
\alias{points,three-method}
\alias{show,three-method}
\alias{store}
\alias{store,three-method}
\alias{text,three-method}
\alias{three}
\alias{three-class}
\title{three class for preparing three.js scripts}
\usage{
\S4method{show}{three}(object)

\S4method{store}{three}(object, directory = NULL, filePrefix = NULL)

\S4method{light}{three}(x, pointLight = list(x = 0, y = 0, z = 0))

\S4method{lines}{three}(x, from, to, color = "0x000000", lwd = 5)

\S4method{points}{three}(x, coords, size = 5, color = "0xcccccc")

\S4method{text}{three}(x, .data, color = "0x000000", size = 15,
  offset = c(x = 5, y = 5, z = 5))

three(type = "base", bgColor = "0x000000", jsUrlPrefix = NULL,
  adjust = list())
}
\arguments{
\item{object}{a three object}

\item{directory}{evident}

\item{filePrefix}{evident, too}

\item{x}{a three object}

\item{pointLight}{a list with the x,y,z-coordinates of a point light}

\item{from}{data frame with coordinates of starting points of lines}

\item{to}{data frame with coordinates of end points of lines}

\item{color}{color of line / point etc.}

\item{lwd}{line width, defaults to 5}

\item{coords}{coordinates of points}

\item{size}{point size}

\item{.data}{a data.frame with three columns providing x/y/z coordinates, and rownames providing text}

\item{offset}{offset of the text}

\item{type}{character vector providing display mode ("base", "anaglyph", "stereo")}

\item{bgColor}{the background color}

\item{jsUrlPrefix}{if not provided, the js files in the package will be linked, if provided
the respective web folder will be referenced}

\item{adjust}{named list, names are replaced by value}
}
\description{
to be displayed with a browser

create a three-object
}
\section{Slots}{

\describe{
\item{\code{js}}{character, the JavaScript code}

\item{\code{json}}{a list with character strings}

\item{\code{type}}{the display mode}

\item{\code{threejsDir}}{directory with js files, may be a web folder}

\item{\code{htmlDir}}{directory where to put the files}
}}
\examples{
\dontrun{
vis <- three(type="base", bgColor="0xcccccc")
vis <- light(vis)
coords <- mtcars[, c("mpg", "wt", "cyl")]
coords <- coords * 10
colnames(coords) <- c("x", "y", "z")
vis <- points(vis, coords=coords, color="#1111FF")
vis <- three:::text(vis, .data=data.frame(coords, row.names=rownames(mtcars)), color="0xff111111")
vis
}
}

